GEFP-26-SA: Fixat så man inte kan ta bort alla spel och att man inte kan uppdatera ett spelnamn till tomt. Börjat på att visa ett spels spelare och macther. Ändra lite grafiskt så det ser bättre ut

+++ src/main/java/com/example/piedpiperdb/DAO/JavaFXActions/GameActions.java
@@ -1,11 +1,12 @@
 package com.example.piedpiperdb.DAO.JavaFXActions;
 
 import com.example.piedpiperdb.DAO.GameDAO;
+import com.example.piedpiperdb.DAO.MatchDAO;
+import com.example.piedpiperdb.DAO.PlayerDAO;
 import com.example.piedpiperdb.Entities.Game;
-import com.example.piedpiperdb.View.AbstractScene;
-import com.example.piedpiperdb.View.GameView;
-import com.example.piedpiperdb.View.HelloApplication;
-import com.example.piedpiperdb.View.StartPage;
+import com.example.piedpiperdb.Entities.Match;
+import com.example.piedpiperdb.Entities.Player;
+import com.example.piedpiperdb.View.*;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.scene.Scene;
@@ -13,7 +14,9 @@ import javafx.scene.control.*;
 import javafx.scene.layout.VBox;
 import javafx.stage.Stage;
 
+import java.util.ArrayList;
 import java.util.HashSet;
+import java.util.List;
 import java.util.Set;
 
 //GEFP-22-SA
@@ -69,22 +72,25 @@ public class GameActions {
 
     //GEFP-22-SA
     public static void deleteGame(ListView gameListView) {
-        Set<Integer>gameById = stringToId(gameListView);
+        if(gameListView.getSelectionModel().getSelectedItems().size() == gameListView.getItems().size()) {
+            AlertBox.displayAlertBox("Delete game","You can't delete all games");
+        }else {
 
-        for(Integer gameId : gameById){
-            System.out.println(gameId);
+            Set<Integer> gameById = stringToId(gameListView);
+
+            for (Integer gameId : gameById) {
+                System.out.println(gameId);
+            }
+
+
+            for (Integer gameId : gameById) {
+                System.out.println("Deleting game with id " + gameId);
+                gameDAO.updatePlayersBeforeDelete(gameId);
+                gameDAO.updateMatchesBeforeDelete(gameId);
+                gameDAO.updateTeamsBeforeDelete(gameId);
+                gameDAO.deleteGameById(gameId);
+            }
         }
-
-
-        for(Integer gameId : gameById){
-            System.out.println("Deleting game with id " + gameId);
-            System.out.println("a-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------");
-            gameDAO.updatePlayersBeforeDelete(gameId);
-            gameDAO.updateMatchesBeforeDelete(gameId);
-            gameDAO.updateTeamsBeforeDelete(gameId);
-            gameDAO.deleteGameById(gameId);
-        }
-
 
     }
     //GEFP-22-SA
@@ -97,34 +103,45 @@ public class GameActions {
     //GEFP-22-SA
     public static void updateGame(TextField textField,ChoiceBox<String> choiceBox) {
 
-        String selectedGame = choiceBox.getValue();
-        System.out.println(selectedGame);
-        System.out.println("j------------------------------------------------------------------------------------------------");
+        //GEFP-26-SA, la till if-sats för om textfield är tomt
+        if(textField.getText().isEmpty() || textField.getText().replaceAll("\\s+","").equals("")){
+            AlertBox.displayAlertBox("Update game","You can't update \""+choiceBox.getValue()+"\" to an empty game name");
+        } else {
+            String selectedGame = choiceBox.getValue();
+            System.out.println(selectedGame);
 
-        String newGameName = textField.getText();
-        System.out.println(newGameName);
-        System.out.println("l----------------------------------------------------------------------------------------------------------------------");
+            String newGameName = textField.getText();
+            System.out.println(newGameName);
 
-        ObservableList<Game> game = getGames();
+            ObservableList<Game> game = getGames();
 
-        for (int i = 0; i < game.size(); i++) {
-            if(game.get(i).getGameName().equals(selectedGame)){
-                Game oldGame = game.get(i);
-                System.out.println(oldGame.getGameName());
-                System.out.println("k-----------------------------------------------------------------------------------------------------------------");
-                gameDAO.updateGame(oldGame, newGameName);
+            for (int i = 0; i < game.size(); i++) {
+                if (game.get(i).getGameName().equals(selectedGame)) {
+                    Game oldGame = game.get(i);
+                    System.out.println(oldGame.getGameName());
+                    gameDAO.updateGame(oldGame, newGameName);
+                }
             }
+
         }
+    }
 
+    public static List<Player> getAllPlayers(){
+        PlayerDAO playerDAO = new PlayerDAO();
+        ObservableList<Player>players = FXCollections.observableArrayList(playerDAO.getAllPlayers());
 
-        /*
-        Game oldGame = gameDAO.getGameById(5);
-        System.out.println("e---------------------------------------------------------------------------------------------------------------");
-
-        gameDAO.updateGame(oldGame, newGameName);*/
-
-
-
+        /*for(Game game : gameDAO.getAllGames()){
+            players.add(game.getPlayers().get(0));
+        }*/
+        return players;
+    }
+    public static List<Match> getAllMatches(){
+        MatchDAO matchDAO = new MatchDAO();
+        ObservableList<Match>matches = FXCollections.observableArrayList(matchDAO.getAllMatches());
+        /*for(Game game : gameDAO.getAllGames()){
+            matches.add(game.getMatches().get(0));
+        }*/
+        return matches;
     }
